Added support for distros using apt-get
Added support to the Dockerfile for distributions using apt-get. The
Dockerfile has been parametised to support override of the base image
and selection of the package manager (and package names) which the image
should use when building.

With this, two additional supported distributions have been added and
documented in the README with build instructions: Debian and Ubuntu.

Cleanup of the container has been added after the package installation
to remove cached files created by the package managers.